# backend/src/main/java/com/legaldev/mvp/web/dto/ComplianceFinding.java
package com.legaldev.mvp.web.dto;

import com.fasterxml.jackson.annotation.JsonInclude;

@JsonInclude(JsonInclude.Include.NON_NULL)
public record ComplianceFinding(
    String ruleId,
    String severity,
    String message,
    Integer lineHint,
    String snippet) {}